requirements.txt
azure-search-documents
azure-identity
python-dotenv
websockets
audio-recorder-streamlit
